refactor(frontend): Replace old page with rewrited one
